moves populatesearch to external file, adds 0.11 migration file
